refactor: move TestConnectionHandler to conftest.py
- Move test helper class to conftest.py
- Fix pytest collection warning
- Update test_logging.py to import from conftest
