+++ tests/integration/test_logging.py
@@ -9,50 +9,7 @@ from mcp.types import LoggingMessageNotification
 
 from mcp_dbutils.base import ConnectionHandler, ConnectionServer
 from mcp_dbutils.log import create_logger
-
-@pytest.mark.no_collect
-class TestConnectionHandler(ConnectionHandler):
-    """Test implementation of ConnectionHandler"""
-    
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # 禁用自动记录统计信息
-        self.stats.record_connection_start = MagicMock()
-        self.stats.record_connection_end = MagicMock()
-    
-    @property
-    def db_type(self) -> str:
-        return "test"
-        
-    async def get_tables(self):
-        return []
-        
-    async def get_schema(self, table_name: str):
-        return ""
-        
-    async def _execute_query(self, sql: str):
-        return ""
-        
-    async def get_table_description(self, table_name: str):
-        return ""
-        
-    async def get_table_ddl(self, table_name: str):
-        return ""
-        
-    async def get_table_indexes(self, table_name: str):
-        return ""
-        
-    async def get_table_stats(self, table_name: str):
-        return ""
-        
-    async def get_table_constraints(self, table_name: str):
-        return ""
-        
-    async def explain_query(self, sql: str):
-        return ""
-        
-    async def cleanup(self):
-        pass
+from .conftest import TestConnectionHandler
 
 @pytest.mark.asyncio
 async def test_handler_logging():
